Move job presenter into web namespace; fix file loading
