+++ lib/sidekiq_adhoc_job.rb
@@ -3,6 +3,7 @@ require 'sidekiq/web'
 
 require 'sidekiq_adhoc_job/utils/string'
 require 'sidekiq_adhoc_job/worker_files_loader'
+require 'sidekiq_adhoc_job/web/job_presenter'
 require 'sidekiq_adhoc_job/web'
 
 module SidekiqAdhocJob
